Reorganize pool blueprints for versioning
