+++ radix-engine-interface/src/blueprints/package/substates.rs
@@ -11,7 +11,7 @@ pub const TRANSACTION_PROCESSOR_CODE_ID: u64 = 7u64;
 pub const METADATA_CODE_ID: u64 = 10u64;
 pub const ROYALTY_CODE_ID: u64 = 11u64;
 pub const ROLE_ASSIGNMENT_CODE_ID: u64 = 12u64;
-pub const POOL_CODE_ID: u64 = 13u64;
+pub const POOL_V1_0_CODE_ID: u64 = 13u64;
 pub const TRANSACTION_TRACKER_CODE_ID: u64 = 14u64;
 pub const TEST_UTILS_CODE_ID: u64 = 15u64;
 pub const CONSENSUS_MANAGER_SECONDS_PRECISION_CODE_ID: u64 = 16u64;
