+++ radix-engine-interface/src/blueprints/resource/mod.rs
@@ -62,7 +62,7 @@ macro_rules! resource_roles {
             pub $updater_field: T,
         }
 
-        impl $roles_struct<RoleDefinition> {
+        impl $roles_struct<$crate::api::node_modules::auth::RoleDefinition> {
             pub fn to_role_init(self) -> $crate::blueprints::resource::RoleAssignmentInit {
                 let mut roles = $crate::blueprints::resource::RoleAssignmentInit::new();
                 roles.define_role($actor_field_name, self.$actor_field);
@@ -71,7 +71,7 @@ macro_rules! resource_roles {
             }
         }
 
-        impl Default for $roles_struct<RoleDefinition> {
+        impl Default for $roles_struct<$crate::api::node_modules::auth::RoleDefinition> {
             fn default() -> Self {
                 Self {
                     $actor_field: Some($default_rule),
