+++ radix-engine-tests/tests/blueprints/pool_two_resource.rs
@@ -1,4 +1,6 @@
-use radix_engine::blueprints::pool::two_resource_pool::*;
+use radix_engine::blueprints::pool::v1::constants::*;
+use radix_engine::blueprints::pool::v1::errors::two_resource_pool::Error as TwoResourcePoolError;
+use radix_engine::blueprints::pool::v1::events::two_resource_pool::*;
 use radix_engine::errors::{ApplicationError, RuntimeError, SystemError, SystemModuleError};
 use radix_engine::transaction::{BalanceChange, TransactionReceipt};
 use radix_engine::types::*;
