+++ radix-engine-tests/tests/system/typed_substate_layout.rs
@@ -240,7 +240,7 @@ fn typed_native_event_type_contains_all_native_events() {
         "Account" => ACCOUNT_PACKAGE_DEFINITION.deref(),
         "Identity" => IDENTITY_PACKAGE_DEFINITION.deref(),
         "AccessController" => ACCESS_CONTROLLER_PACKAGE_DEFINITION.deref(),
-        "Pool" => POOL_PACKAGE_DEFINITION.deref(),
+        "Pool" => POOL_PACKAGE_DEFINITION_V1_0.deref(),
         "TransactionTracker" => TRANSACTION_TRACKER_PACKAGE_DEFINITION.deref(),
         "Resource" => RESOURCE_PACKAGE_DEFINITION.deref(),
         "Package" => PACKAGE_PACKAGE_DEFINITION.deref(),
