+++ radix-engine/src/blueprints/native_schema.rs
@@ -6,7 +6,7 @@ use crate::blueprints::account::*;
 use crate::blueprints::consensus_manager::*;
 use crate::blueprints::identity::*;
 use crate::blueprints::package::*;
-use crate::blueprints::pool::*;
+use crate::blueprints::pool::v1::package::*;
 use crate::blueprints::resource::*;
 use crate::blueprints::transaction_processor::*;
 use crate::blueprints::transaction_tracker::TransactionTrackerNativePackage;
@@ -23,7 +23,8 @@ lazy_static! {
         IdentityNativePackage::definition();
     pub static ref ACCESS_CONTROLLER_PACKAGE_DEFINITION: PackageDefinition =
         AccessControllerNativePackage::definition();
-    pub static ref POOL_PACKAGE_DEFINITION: PackageDefinition = PoolNativePackage::definition();
+    pub static ref POOL_PACKAGE_DEFINITION_V1_0: PackageDefinition =
+        PoolNativePackage::definition(PoolV1MinorVersion::Zero);
     pub static ref TRANSACTION_TRACKER_PACKAGE_DEFINITION: PackageDefinition =
         TransactionTrackerNativePackage::definition();
     pub static ref RESOURCE_PACKAGE_DEFINITION: PackageDefinition =
