+++ radix-engine/src/blueprints/pool/mod.rs
@@ -1,6 +1,1 @@
-pub mod multi_resource_pool;
-pub mod one_resource_pool;
-pub mod two_resource_pool;
-
-mod package;
-pub use package::*;
+pub mod v1;
