--- radix-engine/src/blueprints/pool/multi_resource_pool/error.rs
@@ -1,34 +1,0 @@
-use crate::errors::ApplicationError;
-use crate::errors::RuntimeError;
-use radix_engine_common::types::*;
-use radix_engine_common::ScryptoSbor;
-use sbor::prelude::*;
-
-#[derive(Debug, Clone, PartialEq, Eq, ScryptoSbor)]
-pub enum MultiResourcePoolError {
-    NonFungibleResourcesAreNotAccepted {
-        resource_address: ResourceAddress,
-    },
-    NonZeroPoolUnitSupplyButZeroReserves,
-    InvalidPoolUnitResource {
-        expected: ResourceAddress,
-        actual: ResourceAddress,
-    },
-    ResourceDoesNotBelongToPool {
-        resource_address: ResourceAddress,
-    },
-    MissingOrEmptyBuckets {
-        resource_addresses: IndexSet<ResourceAddress>,
-    },
-    PoolCreationWithSameResource,
-    ContributionOfEmptyBucketError,
-    CantCreatePoolWithLessThanOneResource,
-    DecimalOverflowError,
-    InvalidGetRedemptionAmount,
-}
-
-impl From<MultiResourcePoolError> for RuntimeError {
-    fn from(error: MultiResourcePoolError) -> Self {
-        Self::ApplicationError(ApplicationError::MultiResourcePoolError(error))
-    }
-}
