--- radix-engine/src/blueprints/pool/multi_resource_pool/mod.rs
@@ -1,9 +1,0 @@
-mod blueprint;
-mod error;
-mod events;
-mod substates;
-
-pub use blueprint::*;
-pub use error::*;
-pub use events::*;
-pub use substates::*;
