+++ radix-engine/src/blueprints/pool/v1/constants.rs
@@ -1,0 +1,7 @@
+// Blueprint Idents
+pub const ONE_RESOURCE_POOL_BLUEPRINT_IDENT: &str = "OneResourcePool";
+pub const TWO_RESOURCE_POOL_BLUEPRINT_IDENT: &str = "TwoResourcePool";
+pub const MULTI_RESOURCE_POOL_BLUEPRINT_IDENT: &str = "MultiResourcePool";
+
+// Roles
+pub const POOL_MANAGER_ROLE: &str = "pool_manager_role";
