+++ radix-engine/src/blueprints/pool/v1/v1_0/mod.rs
@@ -1,0 +1,7 @@
+mod multi_resource_pool_blueprint;
+mod one_resource_pool_blueprint;
+mod two_resource_pool_blueprint;
+
+pub use multi_resource_pool_blueprint::*;
+pub use one_resource_pool_blueprint::*;
+pub use two_resource_pool_blueprint::*;
