+++ radix-engine/src/blueprints/pool/v1/v1_0/two_resource_pool_blueprint.rs
@@ -1,221 +1,21 @@
-use crate::blueprints::pool::two_resource_pool::*;
-use crate::blueprints::pool::POOL_MANAGER_ROLE;
-use crate::errors::*;
-use crate::internal_prelude::declare_native_blueprint_state;
+use crate::blueprints::pool::v1::constants::*;
+use crate::blueprints::pool::v1::errors::two_resource_pool::*;
+use crate::blueprints::pool::v1::events::two_resource_pool::*;
+use crate::blueprints::pool::v1::substates::two_resource_pool::*;
 use crate::internal_prelude::*;
 use crate::kernel::kernel_api::*;
-use crate::prelude::BlueprintSchemaInit;
-use crate::types::{ReceiverInfo, TypeRef};
-use crate::{event_schema, roles_template};
 use native_sdk::modules::metadata::*;
 use native_sdk::modules::role_assignment::*;
 use native_sdk::modules::royalty::*;
 use native_sdk::resource::*;
-use native_sdk::runtime::Runtime;
-use radix_engine_common::math::*;
-use radix_engine_common::prelude::*;
-use radix_engine_interface::api::node_modules::auth::RoleDefinition;
-use radix_engine_interface::api::node_modules::auth::ToRoleEntry;
-use radix_engine_interface::api::*;
-use radix_engine_interface::blueprints::component::Global;
-use radix_engine_interface::blueprints::package::{
-    AuthConfig, BlueprintDefinitionInit, BlueprintType, FunctionAuth, MethodAuthTemplate,
-};
+use native_sdk::runtime::*;
+use radix_engine_interface::blueprints::component::*;
 use radix_engine_interface::blueprints::pool::*;
-use radix_engine_interface::blueprints::resource::*;
-use radix_engine_interface::prelude::{
-    BlueprintFunctionsSchemaInit, BlueprintHooksInit, BlueprintStateSchemaInit, FunctionSchemaInit,
-};
-use radix_engine_interface::types::*;
+use radix_engine_interface::prelude::*;
 use radix_engine_interface::*;
-
-pub const TWO_RESOURCE_POOL_BLUEPRINT_IDENT: &'static str = "TwoResourcePool";
-
-declare_native_blueprint_state! {
-    blueprint_ident: TwoResourcePool,
-    blueprint_snake_case: two_resource_pool,
-    features: {
-    },
-    fields: {
-        state:  {
-            ident: State,
-            field_type: {
-                kind: StaticSingleVersioned,
-            },
-            condition: Condition::Always,
-        }
-    },
-    collections: {
-    }
-}
-
-pub type TwoResourcePoolStateV1 = TwoResourcePoolSubstate;
 
 pub struct TwoResourcePoolBlueprint;
 impl TwoResourcePoolBlueprint {
-    pub fn definition() -> BlueprintDefinitionInit {
-        let mut aggregator = TypeAggregator::<ScryptoCustomTypeKind>::new();
-        let feature_set = TwoResourcePoolFeatureSet::all_features();
-        let state = TwoResourcePoolStateSchemaInit::create_schema_init(&mut aggregator);
-
-        let mut functions = index_map_new();
-
-        functions.insert(
-            TWO_RESOURCE_POOL_INSTANTIATE_IDENT.to_string(),
-            FunctionSchemaInit {
-                receiver: None,
-                input: TypeRef::Static(
-                    aggregator.add_child_type_and_descendents::<TwoResourcePoolInstantiateInput>(),
-                ),
-                output: TypeRef::Static(
-                    aggregator.add_child_type_and_descendents::<TwoResourcePoolInstantiateOutput>(),
-                ),
-                export: TWO_RESOURCE_POOL_INSTANTIATE_EXPORT_NAME.to_string(),
-            },
-        );
-
-        functions.insert(
-            TWO_RESOURCE_POOL_CONTRIBUTE_IDENT.to_string(),
-            FunctionSchemaInit {
-                receiver: Some(ReceiverInfo::normal_ref_mut()),
-                input: TypeRef::Static(
-                    aggregator.add_child_type_and_descendents::<TwoResourcePoolContributeInput>(),
-                ),
-                output: TypeRef::Static(
-                    aggregator.add_child_type_and_descendents::<TwoResourcePoolContributeOutput>(),
-                ),
-                export: TWO_RESOURCE_POOL_CONTRIBUTE_EXPORT_NAME.to_string(),
-            },
-        );
-
-        functions.insert(
-            TWO_RESOURCE_POOL_REDEEM_IDENT.to_string(),
-            FunctionSchemaInit {
-                receiver: Some(ReceiverInfo::normal_ref_mut()),
-                input: TypeRef::Static(
-                    aggregator.add_child_type_and_descendents::<TwoResourcePoolRedeemInput>(),
-                ),
-                output: TypeRef::Static(
-                    aggregator.add_child_type_and_descendents::<TwoResourcePoolRedeemOutput>(),
-                ),
-                export: TWO_RESOURCE_POOL_REDEEM_EXPORT_NAME.to_string(),
-            },
-        );
-
-        functions.insert(
-            TWO_RESOURCE_POOL_PROTECTED_DEPOSIT_IDENT.to_string(),
-            FunctionSchemaInit {
-                receiver: Some(ReceiverInfo::normal_ref_mut()),
-                input: TypeRef::Static(
-                    aggregator
-                        .add_child_type_and_descendents::<TwoResourcePoolProtectedDepositInput>(),
-                ),
-                output: TypeRef::Static(
-                    aggregator
-                        .add_child_type_and_descendents::<TwoResourcePoolProtectedDepositOutput>(),
-                ),
-                export: TWO_RESOURCE_POOL_PROTECTED_DEPOSIT_EXPORT_NAME.to_string(),
-            },
-        );
-
-        functions.insert(
-            TWO_RESOURCE_POOL_PROTECTED_WITHDRAW_IDENT.to_string(),
-            FunctionSchemaInit {
-                receiver: Some(ReceiverInfo::normal_ref_mut()),
-                input: TypeRef::Static(
-                    aggregator
-                        .add_child_type_and_descendents::<TwoResourcePoolProtectedWithdrawInput>(),
-                ),
-                output: TypeRef::Static(
-                    aggregator
-                        .add_child_type_and_descendents::<TwoResourcePoolProtectedWithdrawOutput>(),
-                ),
-                export: TWO_RESOURCE_POOL_PROTECTED_WITHDRAW_EXPORT_NAME.to_string(),
-            },
-        );
-
-        functions.insert(
-            TWO_RESOURCE_POOL_GET_REDEMPTION_VALUE_IDENT.to_string(),
-            FunctionSchemaInit {
-                receiver: Some(ReceiverInfo::normal_ref()),
-                input: TypeRef::Static(
-                    aggregator
-                        .add_child_type_and_descendents::<TwoResourcePoolGetRedemptionValueInput>(),
-                ),
-                output: TypeRef::Static(
-                    aggregator
-                        .add_child_type_and_descendents::<TwoResourcePoolGetRedemptionValueOutput>(
-                        ),
-                ),
-                export: TWO_RESOURCE_POOL_GET_REDEMPTION_VALUE_EXPORT_NAME.to_string(),
-            },
-        );
-
-        functions.insert(
-            TWO_RESOURCE_POOL_GET_VAULT_AMOUNTS_IDENT.to_string(),
-            FunctionSchemaInit {
-                receiver: Some(ReceiverInfo::normal_ref()),
-                input: TypeRef::Static(
-                    aggregator
-                        .add_child_type_and_descendents::<TwoResourcePoolGetVaultAmountsInput>(),
-                ),
-                output: TypeRef::Static(
-                    aggregator
-                        .add_child_type_and_descendents::<TwoResourcePoolGetVaultAmountsOutput>(),
-                ),
-                export: TWO_RESOURCE_POOL_GET_VAULT_AMOUNTS_EXPORT_NAME.to_string(),
-            },
-        );
-
-        let event_schema = event_schema! {
-            aggregator,
-            [
-                ContributionEvent,
-                RedemptionEvent,
-                WithdrawEvent,
-                DepositEvent
-            ]
-        };
-
-        let schema = generate_full_schema(aggregator);
-
-        BlueprintDefinitionInit {
-            blueprint_type: BlueprintType::default(),
-            is_transient: false,
-            dependencies: indexset!(),
-            feature_set,
-
-            schema: BlueprintSchemaInit {
-                generics: vec![],
-                schema,
-                state,
-                events: event_schema,
-                types: BlueprintTypeSchemaInit::default(),
-                functions: BlueprintFunctionsSchemaInit { functions },
-                hooks: BlueprintHooksInit::default(),
-            },
-
-            royalty_config: PackageRoyaltyConfig::default(),
-            auth_config: AuthConfig {
-                function_auth: FunctionAuth::AllowAll,
-                method_auth: MethodAuthTemplate::StaticRoleDefinition(roles_template! {
-                    roles {
-                        POOL_MANAGER_ROLE;
-                    },
-                    methods {
-                        // Main Module rules
-                        TWO_RESOURCE_POOL_REDEEM_IDENT => MethodAccessibility::Public;
-                        TWO_RESOURCE_POOL_GET_REDEMPTION_VALUE_IDENT => MethodAccessibility::Public;
-                        TWO_RESOURCE_POOL_GET_VAULT_AMOUNTS_IDENT => MethodAccessibility::Public;
-                        TWO_RESOURCE_POOL_CONTRIBUTE_IDENT => [POOL_MANAGER_ROLE];
-                        TWO_RESOURCE_POOL_PROTECTED_DEPOSIT_IDENT => [POOL_MANAGER_ROLE];
-                        TWO_RESOURCE_POOL_PROTECTED_WITHDRAW_IDENT => [POOL_MANAGER_ROLE];
-                    }
-                }),
-            },
-        }
-    }
-
     pub fn instantiate<Y>(
         (resource_address1, resource_address2): (ResourceAddress, ResourceAddress),
         owner_role: OwnerRole,
@@ -228,7 +28,7 @@ impl TwoResourcePoolBlueprint {
     {
         // A pool can't be created between the same resources - error out if it's
         if resource_address1 == resource_address2 {
-            return Err(TwoResourcePoolError::PoolCreationWithSameResource.into());
+            return Err(Error::PoolCreationWithSameResource.into());
         }
 
         // A pool can't be created where one of the resources is non-fungible - error out if any of
@@ -236,10 +36,7 @@ impl TwoResourcePoolBlueprint {
         for resource_address in [resource_address1, resource_address2] {
             let resource_manager = ResourceManager(resource_address);
             if let ResourceType::NonFungible { .. } = resource_manager.resource_type(api)? {
-                return Err(TwoResourcePoolError::NonFungibleResourcesAreNotAccepted {
-                    resource_address,
-                }
-                .into());
+                return Err(Error::NonFungibleResourcesAreNotAccepted { resource_address }.into());
             }
         }
 
@@ -308,7 +105,7 @@ impl TwoResourcePoolBlueprint {
         )?;
         let royalty = ComponentRoyalty::create(ComponentRoyaltyConfig::default(), api)?;
         let object_id = {
-            let substate = TwoResourcePoolSubstate {
+            let substate = Substate {
                 vaults: [
                     (resource_address1, Vault::create(resource_address1, api)?),
                     (resource_address2, Vault::create(resource_address2, api)?),
@@ -381,13 +178,13 @@ impl TwoResourcePoolBlueprint {
 
             // Ensure that the two buckets given as arguments match the two vaults that the pool has.
             if bucket1_resource_address != vault1_resource_address {
-                return Err(TwoResourcePoolError::ResourceDoesNotBelongToPool {
+                return Err(Error::ResourceDoesNotBelongToPool {
                     resource_address: bucket1_resource_address,
                 }
                 .into());
             }
             if bucket2_resource_address != vault2_resource_address {
-                return Err(TwoResourcePoolError::ResourceDoesNotBelongToPool {
+                return Err(Error::ResourceDoesNotBelongToPool {
                     resource_address: bucket2_resource_address,
                 }
                 .into());
@@ -429,7 +226,7 @@ impl TwoResourcePoolBlueprint {
             })?;
 
             if contribution1 == Decimal::ZERO || contribution2 == Decimal::ZERO {
-                return Err(TwoResourcePoolError::ContributionOfEmptyBucketError.into());
+                return Err(Error::ContributionOfEmptyBucketError.into());
             }
 
             match (
@@ -456,7 +253,7 @@ impl TwoResourcePoolBlueprint {
                         })
                         .and_then(|d| d.checked_round(19, RoundingMode::ToPositiveInfinity))
                         .and_then(|d| Decimal::try_from(d).ok())
-                        .ok_or(TwoResourcePoolError::DecimalOverflowError)?,
+                        .ok_or(Error::DecimalOverflowError)?,
                     contribution1,
                     contribution2,
                 )),
@@ -481,7 +278,7 @@ impl TwoResourcePoolBlueprint {
                         })
                         .and_then(|d| d.checked_round(19, RoundingMode::ToPositiveInfinity))
                         .and_then(|d| Decimal::try_from(d).ok())
-                        .ok_or(TwoResourcePoolError::DecimalOverflowError)?,
+                        .ok_or(Error::DecimalOverflowError)?,
                     contribution1,
                     contribution2,
                 )),
@@ -521,9 +318,9 @@ impl TwoResourcePoolBlueprint {
                     .map(|(c1, c2)| -> Result<(Decimal, Decimal), RuntimeError> {
                         Ok((
                             c1.checked_round(divisibility1, RoundingMode::ToNegativeInfinity)
-                                .ok_or(TwoResourcePoolError::DecimalOverflowError)?,
+                                .ok_or(Error::DecimalOverflowError)?,
                             c2.checked_round(divisibility2, RoundingMode::ToNegativeInfinity)
-                                .ok_or(TwoResourcePoolError::DecimalOverflowError)?,
+                                .ok_or(Error::DecimalOverflowError)?,
                         ))
                     })
                     .filter_map(Result::ok)
@@ -532,15 +329,15 @@ impl TwoResourcePoolBlueprint {
                             let pool_units_to_mint = c1
                                 .checked_div(reserves1)
                                 .and_then(|d| d.checked_mul(pool_unit_total_supply))
-                                .ok_or(TwoResourcePoolError::DecimalOverflowError)?;
+                                .ok_or(Error::DecimalOverflowError)?;
                             Ok((pool_units_to_mint, c1, c2))
                         },
                     )
                     .filter_map(Result::ok)
                     .max_by(|(mint1, _, _), (mint2, _, _)| mint1.cmp(mint2))
-                    .ok_or(TwoResourcePoolError::DecimalOverflowError)
+                    .ok_or(Error::DecimalOverflowError)
                 }
-                (true, _, _) => Err(TwoResourcePoolError::NonZeroPoolUnitSupplyButZeroReserves),
+                (true, _, _) => Err(Error::NonZeroPoolUnitSupplyButZeroReserves),
             }
         }?;
 
@@ -599,7 +396,7 @@ impl TwoResourcePoolBlueprint {
         // Ensure that the passed pool resources are indeed pool resources
         let bucket_resource_address = bucket.resource_address(api)?;
         if bucket_resource_address != substate.pool_unit_resource_manager.0 {
-            return Err(TwoResourcePoolError::InvalidPoolUnitResource {
+            return Err(Error::InvalidPoolUnitResource {
                 expected: substate.pool_unit_resource_manager.0,
                 actual: bucket_resource_address,
             }
@@ -679,7 +476,7 @@ impl TwoResourcePoolBlueprint {
             Runtime::emit_event(api, event)?;
             Ok(())
         } else {
-            Err(TwoResourcePoolError::ResourceDoesNotBelongToPool { resource_address }.into())
+            Err(Error::ResourceDoesNotBelongToPool { resource_address }.into())
         }
     }
 
@@ -710,7 +507,7 @@ impl TwoResourcePoolBlueprint {
 
             Ok(bucket)
         } else {
-            Err(TwoResourcePoolError::ResourceDoesNotBelongToPool { resource_address }.into())
+            Err(Error::ResourceDoesNotBelongToPool { resource_address }.into())
         }
     }
 
@@ -733,7 +530,7 @@ impl TwoResourcePoolBlueprint {
             || amount_of_pool_units.is_zero()
             || amount_of_pool_units > pool_units_total_supply
         {
-            return Err(TwoResourcePoolError::InvalidGetRedemptionAmount.into());
+            return Err(Error::InvalidGetRedemptionAmount.into());
         }
 
         let mut reserves = index_map_new();
@@ -792,7 +589,7 @@ impl TwoResourcePoolBlueprint {
     fn lock_and_read<Y>(
         api: &mut Y,
         lock_flags: LockFlags,
-    ) -> Result<(TwoResourcePoolSubstate, SubstateHandle), RuntimeError>
+    ) -> Result<(Substate, SubstateHandle), RuntimeError>
     where
         Y: ClientApi<RuntimeError>,
     {
@@ -827,14 +624,14 @@ impl TwoResourcePoolBlueprint {
                     let amount_owed = pool_units_to_redeem
                         .checked_div(pool_units_total_supply)
                         .and_then(|d| d.checked_mul(reserves))
-                        .ok_or(TwoResourcePoolError::DecimalOverflowError)?;
+                        .ok_or(Error::DecimalOverflowError)?;
 
                     let amount_owed = if divisibility == 18 {
                         amount_owed
                     } else {
                         amount_owed
                             .checked_round(divisibility, RoundingMode::ToNegativeInfinity)
-                            .ok_or(TwoResourcePoolError::DecimalOverflowError)?
+                            .ok_or(Error::DecimalOverflowError)?
                     };
 
                     Ok((resource_address, amount_owed))
