+++ scrypto-test/src/prelude.rs
@@ -8,9 +8,6 @@ pub use radix_engine::blueprints::identity::*;
 pub use radix_engine::blueprints::models::*;
 pub use radix_engine::blueprints::native_schema::*;
 pub use radix_engine::blueprints::package::*;
-pub use radix_engine::blueprints::pool::multi_resource_pool::*;
-pub use radix_engine::blueprints::pool::one_resource_pool::*;
-pub use radix_engine::blueprints::pool::two_resource_pool::*;
 pub use radix_engine::blueprints::pool::*;
 pub use radix_engine::blueprints::resource::*;
 pub use radix_engine::blueprints::transaction_processor::*;
